+++ pkg/parsers/parsers_test.go
@@ -5,33 +5,6 @@ import (
 	"testing"
 )
 
-func TestParseKeyValueOpt(t *testing.T) {
-	invalids := map[string]string{
-		"":    "unable to parse key/value option: ",
-		"key": "unable to parse key/value option: key",
-	}
-	for invalid, expectedError := range invalids {
-		if _, _, err := ParseKeyValueOpt(invalid); err == nil || err.Error() != expectedError {
-			t.Fatalf("Expected error %v for %v, got %v", expectedError, invalid, err)
-		}
-	}
-	valids := map[string][]string{
-		"key=value":               {"key", "value"},
-		" key = value ":           {"key", "value"},
-		"key=value1=value2":       {"key", "value1=value2"},
-		" key = value1 = value2 ": {"key", "value1 = value2"},
-	}
-	for valid, expectedKeyValue := range valids {
-		key, value, err := ParseKeyValueOpt(valid)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if key != expectedKeyValue[0] || value != expectedKeyValue[1] {
-			t.Fatalf("Expected {%v: %v} got {%v: %v}", expectedKeyValue[0], expectedKeyValue[1], key, value)
-		}
-	}
-}
-
 func TestParseUintList(t *testing.T) {
 	valids := map[string]map[int]bool{
 		"":             {},
